wificond: Remove supplicant/hostapd start/stop
This is no longer needed since we'll be using the lazy HAL startup
mechanism provided by HIDL to start/stop supplicant.
Created a new method for registering the softap callbacks (previously
was a part of startHostapd. The callback will be removed when we
teardown the interface.

libwifi-system (HostapdManager & SupplicantManager) will be removed in
subsequent CL's once VTS tests migrates to the new mechanism.

Bug: 72394251
Bug: 30311493
Test: ./system/connectivity/wificond/runtests.sh
Test: Device boots up & connects to wifi networks.

Change-Id: If41d40af2970d0e09b0d73388fe4f7a0640cd52b