Mark EventLoop::StopWatchFileDescriptor() as pure virtual
Bug: http://b/116873221

If not, Clang assumes that the TU which defines this function will
contain the VTable for the class as well.  Since there's no out-of-line
definition of this funciton, no TU ends up with the VTable.

This causes a problem with coverage builds, which are built with -O0,
where calls don't get inlined, thereby requiring a definition of the
VTable.  For non O0 builds, the VTable is not required since the virtual
calls get inlined/optimized-out.

Test: Build with and without -O0.
Change-Id: Iabc7f7030025286223f03258ce57a0f562528c98