wificond: Fix corruption of iface to wiphy map
In C++, using "map[key]" to read a key from the map results in the
creation of "key" if it does not already exist in the map. Need to use
"map.find(key)"  to check for existence before read. This was causing a
corruption of the |iface_to_wiphy_index_map_| in the dump() function.
This issue has existed for a while in S (introduced as a part of 60Ghz
upstreams). This issue was however exposed by the recent changes added
in ag/13225431 (since we check for the map contents before updating the
|band_to_wiphy_index| map).

Instead store the |wiphy_index| directly in InterfaceInfo to help in
dumping.

Also, removed/reduced some logs previously added for debugging the
issue.

Bug: 177936058
Test: Manual steps which repro'ed the issue before the fix:
i. Toggle wifi off.
ii. adb shell dumpsys wifinl80211 (or trigger bugreport)
iii. Turn wifi on.
iv. Verify scans are not working without the fix & works with the fix.

Change-Id: I84b85b24a6c0a0efc6c22bb74560e5ccc5f0b39d